requirements.txt
pygame==2.1.2
pyserial==3.5
python-dotenv==0.20.0
